ops(ci): four MySQL-only failures, three of them real
The MySQL job is the one that runs the suite against the engine production
uses, and it has been carrying four failures that SQLite cannot see.

**`charter_agreements.template_version` was three characters too short for its
own constant.** `provisional-2026-09` is nineteen and the column is
`varchar(16)`. MySQL answers `SQLSTATE[22001]: Data too long` and no charter
agreement saves at all; SQLite does not enforce a declared length, so six tests
passed locally and the same six failed in CI. Widened to 32 — the
`provisional-` prefix is load-bearing (it keeps every document produced under a
template no lawyer has read distinguishable from one produced under the real
text) so the column moves, not the constant.

**Three audit assertions compared key order.** MySQL's JSON type normalises an
object by sorting its keys shortest-first, so `…_to` comes back before `…_from`
there and after it on SQLite. `toEqual` asserts the contents, which is what the
tests are about; the order is the database's business.

**Two tests assumed the array cache.** `RateLimiter::attempts()` returns `'1'`
from Redis and `1` from the array store, and the idempotency retention test
travels a day in Carbon — which the array and database drivers honour, because
they compare an expiry against `Carbon::now()`, and Redis does not, because it
expires on its own clock. The counter is cast; the retention test skips on
Redis with a sentence saying what is skipped is the simulation of a day passing
and not the promise, which is one line in `IdempotencyStore::ttlSeconds()` and
the same number on every driver.

+++ tests/Feature/Guest/TokenRateLimitTest.php
@@ -88,10 +88,13 @@ it('counts a success against the lookup budget and not the failure one', functio
 
     get('/b/' . $booking->manage_token)->assertOk();
 
-    expect(RateLimiter::attempts(ThrottleTokenLookups::lookupKey('127.0.0.1')))->toBe(1)
+    // Cast, because a counter read back from Redis is the string `'1'` and from
+    // the array store the integer `1`. The number is the assertion; which
+    // driver the cache happens to be is not.
+    expect((int) RateLimiter::attempts(ThrottleTokenLookups::lookupKey('127.0.0.1')))->toBe(1)
         // Counting successes as failures would lock a guest out of their own
         // booking for refreshing it, which is the failure this split avoids.
-        ->and(RateLimiter::attempts(ThrottleTokenLookups::failureKey('127.0.0.1')))->toBe(0);
+        ->and((int) RateLimiter::attempts(ThrottleTokenLookups::failureKey('127.0.0.1')))->toBe(0);
 });
 
 it('limits the voucher page on the same budget, which is the harder one for a short code', function (): void {
